my-line-bot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
line-bot-sdk==3.5.0
aiohttp==3.9.1
gspread==5.12.1
gspread-asyncio==1.3.1
google-auth==2.23.4
python-dotenv==1.0.0
